removed last stdlib dependency

+++ src/ed25519.h
@@ -1,8 +1,10 @@
 #ifndef ED25519_H
 #define ED25519_H
 
-int ed25519_sign(unsigned char *signature, const unsigned char *message, unsigned int message_len, const unsigned char *sign_key);
-int ed25519_verify(const unsigned char *signature, const unsigned char *message, unsigned int message_len, const unsigned char *verify_key);
+#include <stddef.h>
+
+int ed25519_sign(unsigned char *signature, const unsigned char *message, size_t message_len, const unsigned char *sign_key);
+int ed25519_verify(const unsigned char *signature, const unsigned char *message, size_t message_len, const unsigned char *verify_key);
 int ed25519_create_keypair(unsigned char *verify_key, unsigned char *sign_key, unsigned char *seed);
 
 #ifndef ED25519_NO_SEED
